requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
django-cors-headers>=4.5.0
Pillow>=11.0.0
mysqlclient>=2.2.4
python-decouple>=3.8
gunicorn>=23.0.0
whitenoise>=6.8.0
